Day 3: Binary Diagnostic
This time I couldn't do it in one pass. In fact, I had to do one pass for the oxygen rating
and another for the CO2 rating.
PHP supports copy-on-write so I only had to pass the original input to the two functions.
PHP things I discovered : shorthand lamda syntax, generators, nullable types