Day 5: Hydrothermal Venture
Lessons learned: copy-on-write
I had to pass the array by reference to the functions that modify it,
because otherwise the outer reference doesn't get mutated by the
functions. That's because PHP supports a copy-on-write mechanism. By
default, it passes a reference of the array. But as soon as the function
attempts to modify it, the engine makes a copy of that array, and the
function operates on the copy, leaving the original one intact.